move invoke span annotation into function scheduled hook
